Clean-up dk2 > transport.c
