+++ omi/firmware/firmware/omi2/omi2.conf
@@ -139,14 +139,41 @@ CONFIG_AUDIO_DMIC=y
 CONFIG_AUDIO_DMIC_NRFX_PDM=y
 CONFIG_NRFX_PDM0=y  # PDM microphone
 
+
+#
+# Large BLE packets / BLE Buffers 
+#
+# CONFIG_BT_GATT_CLIENT=y
+# CONFIG_BT_THROUGHPUT=y
+
+CONFIG_BT_L2CAP_TX_MTU=498
+
+CONFIG_BT_CTLR_DATA_LENGTH_MAX=251
+CONFIG_BT_CTLR_PHY_2M=y
+CONFIG_BT_CTLR_PHY_CODED=y
+CONFIG_BT_CTLR_SDC_MAX_CONN_EVENT_LEN_DEFAULT=400000
+CONFIG_BT_CONN_TX_MAX=20
+CONFIG_BT_BUF_ACL_RX_SIZE=1024
+CONFIG_BT_L2CAP_TX_BUF_COUNT=10
+CONFIG_BT_BUF_ACL_TX_SIZE=2048
+# CONFIG_BT_ATT_TX_COUNT=10
+# CONFIG_BT_EXT_ADV=y
+# CONFIG_BT_PER_ADV=y
+# CONFIG_BT_OBSERVER=y
+# CONFIG_BT_ATT_PREPARE_COUNT=2
+CONFIG_BT_USER_DATA_LEN_UPDATE=y
+CONFIG_BT_AUTO_DATA_LEN_UPDATE=y
+CONFIG_BT_USER_PHY_UPDATE=y
+CONFIG_BT_AUTO_PHY_UPDATE=y
+
 #
 # Omi features configuration
 #
 CONFIG_OMI_CODEC_OPUS=y
 CONFIG_OMI_ENABLE_OFFLINE_STORAGE=y
-CONFIG_OMI_ENABLE_ACCELEROMETER=y
-CONFIG_OMI_ENABLE_BUTTON=y
-CONFIG_OMI_ENABLE_SPEAKER=y
-CONFIG_OMI_ENABLE_BATTERY=y
-CONFIG_OMI_ENABLE_USB=y
-CONFIG_OMI_ENABLE_HAPTIC=y
+CONFIG_OMI_ENABLE_ACCELEROMETER=n
+CONFIG_OMI_ENABLE_BUTTON=n
+CONFIG_OMI_ENABLE_SPEAKER=n
+CONFIG_OMI_ENABLE_BATTERY=n
+CONFIG_OMI_ENABLE_USB=n
+CONFIG_OMI_ENABLE_HAPTIC=n
