+++ omi/firmware/firmware/omi2/src/lib/dk2/transport.c
@@ -12,15 +12,12 @@
 #include <zephyr/sys/ring_buffer.h>
 #include "transport.h"
 #include "config.h"
-#include "utils.h"
-// #include "nfc.h"
 #include "speaker.h"
 #include "sdcard.h"
 #include "storage.h"
 #include "button.h"
 #include "mic.h"
-#include "lib/battery/battery.h"
-// #include "friend.h"
+#include "accel.h"
 LOG_MODULE_REGISTER(transport, CONFIG_LOG_DEFAULT_LEVEL);
 
 #define MAX_STORAGE_BYTES 0xFFFF0000
@@ -88,133 +85,7 @@ static struct bt_gatt_attr dfu_service_attr[] = {
 };
 
 static struct bt_gatt_service dfu_service = BT_GATT_SERVICE(dfu_service_attr);
-//Acceleration data
-//this code activates the onboard accelerometer. some cute ideas may include shaking the necklace to color strobe
-//
-static struct sensors mega_sensor;
-static struct device *lsm6dsl_dev;
-//Arbritrary uuid, feel free to change
-static struct bt_uuid_128 accel_uuid = BT_UUID_INIT_128(BT_UUID_128_ENCODE(0x32403790,0x0000,0x1000,0x7450,0xBF445E5829A2));
-static struct bt_uuid_128 accel_uuid_x = BT_UUID_INIT_128(BT_UUID_128_ENCODE(0x32403791,0x0000,0x1000,0x7450,0xBF445E5829A2));
 
-static void accel_ccc_config_changed_handler(const struct bt_gatt_attr *attr, uint16_t value);
-static ssize_t accel_data_read_characteristic(struct bt_conn *conn, const struct bt_gatt_attr *attr, void *buf, uint16_t len, uint16_t offset);
-
-static struct bt_gatt_attr accel_service_attr[] = {
-    BT_GATT_PRIMARY_SERVICE(&accel_uuid),//primary description
-    BT_GATT_CHARACTERISTIC(&accel_uuid_x.uuid, BT_GATT_CHRC_READ | BT_GATT_CHRC_NOTIFY, BT_GATT_PERM_READ, accel_data_read_characteristic, NULL, NULL),//data type
-    BT_GATT_CCC(accel_ccc_config_changed_handler, BT_GATT_PERM_READ | BT_GATT_PERM_WRITE),//scheduler
-};
-static struct bt_gatt_service accel_service = BT_GATT_SERVICE(accel_service_attr);
-
-static ssize_t accel_data_read_characteristic(struct bt_conn *conn, const struct bt_gatt_attr *attr, void *buf, uint16_t len, uint16_t offset)
-{
-    LOG_INF("Acceleration data read characteristic");
-    int axis_mode = 6; //3 for accel, 6 for (also) gyro
-    return bt_gatt_attr_read(conn, attr, buf, len, offset, &axis_mode, sizeof(axis_mode));
-}
-
-
-#define ACCEL_REFRESH_INTERVAL 1000 // 1.0 seconds
-
-void broadcast_accel(struct k_work *work_item);
-K_WORK_DELAYABLE_DEFINE(accel_work, broadcast_accel);
-
-void broadcast_accel(struct k_work *work_item) {
-
-    sensor_sample_fetch_chan(lsm6dsl_dev, SENSOR_CHAN_ACCEL_XYZ);
-    sensor_channel_get(lsm6dsl_dev, SENSOR_CHAN_ACCEL_X, &mega_sensor.a_x);
-    sensor_channel_get(lsm6dsl_dev, SENSOR_CHAN_ACCEL_Y, &mega_sensor.a_y);
-    sensor_channel_get(lsm6dsl_dev, SENSOR_CHAN_ACCEL_Z, &mega_sensor.a_z);
-
-    sensor_sample_fetch_chan(lsm6dsl_dev, SENSOR_CHAN_GYRO_XYZ);
-    sensor_channel_get(lsm6dsl_dev, SENSOR_CHAN_GYRO_X, &mega_sensor.g_x);
-    sensor_channel_get(lsm6dsl_dev, SENSOR_CHAN_GYRO_Y, &mega_sensor.g_y);
-    sensor_channel_get(lsm6dsl_dev, SENSOR_CHAN_GYRO_Z, &mega_sensor.g_z);
-
-   //only time mega sensor is changed is through here (hopefully),  so no chance of race condition
-    int err = bt_gatt_notify(current_connection, &accel_service.attrs[1], &mega_sensor, sizeof(mega_sensor));
-    if (err)
-    {
-       LOG_ERR("Error updating Accelerometer data");
-    }
-    k_work_reschedule(&accel_work, K_MSEC(ACCEL_REFRESH_INTERVAL));
-}
-
-struct gpio_dt_spec accel_gpio_pin = {.port = DEVICE_DT_GET(DT_NODELABEL(gpio1)), .pin=8, .dt_flags = GPIO_INT_DISABLE};
-
-//use d4,d5
-static void accel_ccc_config_changed_handler(const struct bt_gatt_attr *attr, uint16_t value)
-{
-    if (value == BT_GATT_CCC_NOTIFY)
-    {
-        LOG_INF("Client subscribed for notifications");
-    }
-    else if (value == 0)
-    {
-        LOG_INF("Client unsubscribed from notifications");
-    }
-    else
-    {
-        LOG_ERR("Invalid CCC value: %u", value);
-    }
-}
-int accel_start()
-{
-    struct sensor_value odr_attr;
-    lsm6dsl_dev = DEVICE_DT_GET_ONE(st_lsm6dsl);
-    k_msleep(50);
-    if (lsm6dsl_dev == NULL)
-    {
-        LOG_ERR("Could not get LSM6DSL device");
-        return 0;
-    }
-    if (!device_is_ready(lsm6dsl_dev))
-    {
-        LOG_ERR("LSM6DSL: not ready");
-        return 0;
-    }
-    odr_attr.val1 = 10;
-    odr_attr.val2 = 0;
-
-
-
-    if (gpio_is_ready_dt(&accel_gpio_pin))
-    {
-        LOG_PRINTK("Speaker Pin ready\n");
-    }
-    else
-    {
-        LOG_PRINTK("Error setting up speaker Pin\n");
-        return -1;
-    }
-    if (gpio_pin_configure_dt(&accel_gpio_pin, GPIO_OUTPUT_INACTIVE) < 0)
-    {
-        LOG_PRINTK("Error setting up Haptic Pin\n");
-        return -1;
-    }
-    gpio_pin_set_dt(&accel_gpio_pin, 1);
-    if (sensor_attr_set(lsm6dsl_dev, SENSOR_CHAN_ACCEL_XYZ,
-        SENSOR_ATTR_SAMPLING_FREQUENCY, &odr_attr) < 0)
-    {
-        LOG_ERR("Cannot set sampling frequency for Accelerometer.");
-        return 0;
-    }
-    if (sensor_attr_set(lsm6dsl_dev, SENSOR_CHAN_GYRO_XYZ,
-        SENSOR_ATTR_SAMPLING_FREQUENCY, &odr_attr) < 0) {
-        LOG_ERR("Cannot set sampling frequency for gyro.");
-        return 0;
-    }
-    if (sensor_sample_fetch(lsm6dsl_dev) < 0)
-    {
-        LOG_ERR("Sensor sample update error");
-        return 0;
-    }
-
-    LOG_INF("Accelerometer is ready for use \n");
-
-    return 1;
-}
 // Advertisement data
 static const struct bt_data bt_ad[] = {
     BT_DATA_BYTES(BT_DATA_FLAGS, (BT_LE_AD_GENERAL | BT_LE_AD_NO_BREDR)),
@@ -294,9 +165,10 @@ static void dfu_ccc_config_changed_handler(const struct bt_gatt_attr *attr, uint
 static ssize_t dfu_control_point_write_handler(struct bt_conn *conn, const struct bt_gatt_attr *attr, const void *buf, uint16_t len, uint16_t offset, uint8_t flags)
 {
     LOG_INF("dfu_control_point_write_handler");
+    uint32_t val = 0xA8;
     if (len == 1 && ((uint8_t *)buf)[0] == 0x06)
     {
-        NRF_POWER->GPREGRET = 0xA8;
+        nrf_power_gpregret_set(NRF_POWER, 0, val);
         NVIC_SystemReset();
     }
     else if (len == 2 && ((uint8_t *)buf)[0] == 0x01)
@@ -304,7 +176,7 @@ static ssize_t dfu_control_point_write_handler(struct bt_conn *conn, const struc
         uint8_t notification_value = 0x10;
         bt_gatt_notify(conn, attr, &notification_value, sizeof(notification_value));
 
-        NRF_POWER->GPREGRET = 0xA8;
+        nrf_power_gpregret_set(NRF_POWER, 0, val);
         NVIC_SystemReset();
     }
     return len;
@@ -813,7 +685,7 @@ int transport_start()
     else
     {
         LOG_INF("Accelerometer initialized");
-        bt_gatt_service_register(&accel_service);
+        register_accel_service(current_connection);
     }
 #endif
     //  Enable button
@@ -895,10 +767,4 @@ int broadcast_audio_packets(uint8_t *buffer, size_t size)
     }
 
     return 0;
-}
-
-
-void accel_off()
-{
-    gpio_pin_set_dt(&accel_gpio_pin, 0);
 }
