+++ omi/firmware/firmware/omi2/src/lib/dk2/transport.h
@@ -2,16 +2,6 @@
 #define TRANSPORT_H
 
 #include <zephyr/drivers/sensor.h>
-typedef struct sensors {
-
-	struct sensor_value a_x;
-	struct sensor_value a_y;
-	struct sensor_value a_z;
-    struct sensor_value g_x;
-    struct sensor_value g_y;
-    struct sensor_value g_z;
-
-};
 /**
  * @brief Initialize the BLE transport logic
  *
@@ -24,6 +14,4 @@ int broadcast_audio_packets(uint8_t *buffer, size_t size);
 struct bt_conn *get_current_connection();
 int bt_on();
 int bt_off();
-
-void accel_off();
 #endif
